Smart Match 3™
I was persuaded by Smylers’ argument concerning string equivalence:
<20120905143040.GZ1742@stripey.com>

But it is not too late for me to be overruled since this is not in
blead yet.

The list is:

      RHS type
   1. undef
   2. ~~ overloading
   3. reftype eq CODE
   4. reftype eq REGEXP
   5. fall back to eq

One question still remains:  Should undef ~~ "string" give a warning?
Currently it doesn’t, but neither does it call eq overloading on the
RHS if we actually have undef ~~ $overloaded_object_without_smartmatch.

This cannot be merged until we solve the fact that smart match
is used in:

cpan/autodie/lib/Fatal.pm
cpan/autodie/t/exceptions.t

This does not deal with loop controls yet, just smartmatch and the
smartmatch aspects of of givwhen.