fix $foo ~~ /re/, autoref and forbid anything not permitted
The "Smart Match 3" patch removed conversion of:

  $foo ~~ /bar/

to:

  $foo ~~ qr/bar/

which this change reinstates.

Also this reinstates auto-reffing of the left-hand argument so that
code like:

  @foo ~~ \&bar

will call bar(\@foo).

Unfortunately this can lead to:

  @foo ~~ "".\@foo

matching, since the left is auto-reffed, then stringified for comparison
with the string on the right.

Additionally, we forbid:

  $foo ~~ @bar       # compile-time
  $foo ~~ %bar       # compile-time
  $foo ~~ @bar[0.1]  # compile-time
  $foo ~~ @bar{0.1}  # compile-time
  $foo ~~ $otherref  # run-time, disallow any unblessed reference not
                     # a code or regexp