Fiddle with starting directory.
t/porting/podcheck.t, invoked without either '--add_link' or '--regen', is
documented to be callable either from the top-level directory:

  ./perl -Ilib t/porting/podcheck.t

... or from t/ via the harness (as in 'make test'):

  cd t;./perl harness porting/podcheck.t;cd -

I've begun to refactor code in podcheck.t into subroutines and to move those
subroutines into t/lib/PodcheckUtils.pm.  But those subroutines, as well as
code still left in podcheck.t, depend on functions require-ed in from
regen/regen_lib.pl.  This poses the problem of locating regen/regen_lib.pl:
Is it in ./ or in ../?  Depends on whether we have chdir-ed to t/ or not.

For the time being, invoking podcheck.t via the harness after having chdir-ed
to t/ is broken.  But the other invocation methods seem to be working.  So
let's continue along the path of encapsulating podcheck.t code into
subroutines and moving them into the module.  We'll come back to the harness
problem later.